# services/apigateway/docs/swagger.yaml
basePath: /
definitions:
...
  events.createEventTypeRequest:
    properties:
      name:
        type: string
      primary_persons_count:
        type: integer
      primary_persons_mode:
        enum:
        - FIXED
        - UNLIMITED
        type: string
    type: object
...
  events.eventSuccessData:
    properties:
      event:
        $ref: '#/definitions/events.eventJSON'
    type: object
  events.eventSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.eventSuccessData'
    type: object
...
  events.eventTypeSuccessData:
    properties:
      event_type:
        $ref: '#/definitions/events.eventTypeJSON'
    type: object
  events.eventTypeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.eventTypeSuccessData'
    type: object
  events.eventTypesSuccessData:
    properties:
      event_types:
        items:
          $ref: '#/definitions/events.eventTypeJSON'
        type: array
    type: object
  events.eventTypesSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.eventTypesSuccessData'
    type: object
  events.eventsStatusData:
    properties:
      status:
        type: string
    type: object
  events.eventsStatusSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.eventsStatusData'
    type: object
  events.eventsSuccessData:
    properties:
      events:
        items:
          $ref: '#/definitions/events.eventJSON'
        type: array
    type: object
  events.eventsSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.eventsSuccessData'
    type: object
  events.upsertEventRequest:
    properties:
      additional_person_ids:
        items:
          type: string
        type: array
      date_bound:
        enum:
        - EXACT
        - NOT_BEFORE
        - NOT_AFTER
        type: string
      date_iso:
        type: string
      date_precision:
        enum:
        - DAY
        - MONTH
        - YEAR
        type: string
      date_unknown:
        type: boolean
      event_type_id:
        type: string
      primary_person_ids:
        items:
          type: string
        type: array
      tree_id:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API gateway for auth, family tree, events, photos and visualisation
    services.
  title: MyForebears API Gateway
  version: "1.0"
paths:
...
  /api/event-types/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventTypesSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List event types
      tags:
      - event-types
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/events.createEventTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventTypeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create event type
      tags:
      - event-types
  /api/event-types/{event_type_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Event type ID
        in: path
        name: event_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventsStatusSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete event type
      tags:
      - event-types
    get:
      consumes:
      - application/json
      parameters:
      - description: Event type ID
        in: path
        name: event_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventTypeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event type
      tags:
      - event-types
  /api/events/{tree_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventsSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List events by tree
      tags:
      - events
    post:
      consumes:
      - application/json
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/events.upsertEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create event
      tags:
      - events
  /api/events/{tree_id}/{event_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventsStatusSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete event
      tags:
      - events
    get:
      consumes:
      - application/json
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event
      tags:
      - events
    put:
      consumes:
      - application/json
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/events.upsertEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.eventSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update event
      tags:
      - events
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"