# services/apigateway/internal/handlers/events/handlers.go
package events

import (
	"encoding/json"
	"log/slog"
	"net/http"
	"strings"

	eventspb "github.com/Bad-Utya/myforebears-backend/gen/go/events"
	eventsclient "github.com/Bad-Utya/myforebears-backend/services/apigateway/internal/clients/events"
	"github.com/Bad-Utya/myforebears-backend/services/apigateway/internal/lib/grpcerr"
	"github.com/Bad-Utya/myforebears-backend/services/apigateway/internal/lib/response"
	"github.com/Bad-Utya/myforebears-backend/services/apigateway/internal/middleware"
	"github.com/go-chi/chi/v5"
)

type Handler struct {
	log    *slog.Logger
	client *eventsclient.Client
}

func New(log *slog.Logger, client *eventsclient.Client) *Handler {
	return &Handler{log: log, client: client}
}

type createEventTypeRequest struct {
	Name                string `json:"name"`
	PrimaryPersonsMode  string `json:"primary_persons_mode" enums:"FIXED,UNLIMITED"`
	PrimaryPersonsCount int    `json:"primary_persons_count"`
}

type upsertEventRequest struct {
	TreeID              string   `json:"tree_id"`
	EventTypeID         string   `json:"event_type_id"`
	PrimaryPersonIDs    []string `json:"primary_person_ids"`
	AdditionalPersonIDs []string `json:"additional_person_ids"`
	DateISO             string   `json:"date_iso"`
	DateUnknown         bool     `json:"date_unknown"`
	DatePrecision       string   `json:"date_precision" enums:"DAY,MONTH,YEAR"`
	DateBound           string   `json:"date_bound" enums:"EXACT,NOT_BEFORE,NOT_AFTER"`
}

type eventTypeJSON struct {
	ID                  string `json:"id"`
	OwnerUserID         int32  `json:"owner_user_id"`
	IsSystem            bool   `json:"is_system"`
	Name                string `json:"name"`
	PrimaryPersonsMode  string `json:"primary_persons_mode" enums:"PRIMARY_PERSONS_MODE_UNSPECIFIED,PRIMARY_PERSONS_MODE_FIXED,PRIMARY_PERSONS_MODE_UNLIMITED"`
	PrimaryPersonsCount int32  `json:"primary_persons_count"`
}

type eventJSON struct {
	ID                  string   `json:"id"`
	TreeID              string   `json:"tree_id"`
	EventTypeID         string   `json:"event_type_id"`
	PrimaryPersonIDs    []string `json:"primary_person_ids"`
	AdditionalPersonIDs []string `json:"additional_person_ids"`
	DateISO             string   `json:"date_iso"`
	DateUnknown         bool     `json:"date_unknown"`
	IsAutogenerated     bool     `json:"is_autogenerated"`
	DatePrecision       string   `json:"date_precision" enums:"EVENT_DATE_PRECISION_UNSPECIFIED,EVENT_DATE_PRECISION_DAY,EVENT_DATE_PRECISION_MONTH,EVENT_DATE_PRECISION_YEAR"`
	DateBound           string   `json:"date_bound" enums:"EVENT_DATE_BOUND_UNSPECIFIED,EVENT_DATE_BOUND_EXACT,EVENT_DATE_BOUND_NOT_BEFORE,EVENT_DATE_BOUND_NOT_AFTER"`
	CreatedAtUnix       int64    `json:"created_at_unix"`
	UpdatedAtUnix       int64    `json:"updated_at_unix"`
}

type eventTypeSuccessData struct {
	EventType eventTypeJSON `json:"event_type"`
}

type eventTypesSuccessData struct {
	EventTypes []eventTypeJSON `json:"event_types"`
}

type eventSuccessData struct {
	Event eventJSON `json:"event"`
}

type eventsSuccessData struct {
	Events []eventJSON `json:"events"`
}

type eventsStatusData struct {
	Status string `json:"status"`
}

type eventTypeSuccessResponse struct {
	Data eventTypeSuccessData `json:"data"`
}

type eventTypesSuccessResponse struct {
	Data eventTypesSuccessData `json:"data"`
}

type eventSuccessResponse struct {
	Data eventSuccessData `json:"data"`
}

type eventsSuccessResponse struct {
	Data eventsSuccessData `json:"data"`
}

type eventsStatusSuccessResponse struct {
	Data eventsStatusData `json:"data"`
}

// CreateEventType creates a custom event type for current user.
// @Summary Create event type
// @Tags event-types
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param request body createEventTypeRequest true "Request body"
// @Success 200 {object} eventTypeSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/event-types/ [post]
func (h *Handler) CreateEventType(w http.ResponseWriter, r *http.Request) {
	userID, err := middleware.UserIDFromContext(r.Context())
	if err != nil {
		response.Error(w, http.StatusUnauthorized, "unauthorized", "invalid token claims")
		return
	}

	var req createEventTypeRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "bad_request", "invalid request body")
		return
	}

	resp, err := h.client.CreateEventType(r.Context(), &eventspb.CreateEventTypeRequest{
		RequestUserId:       int32(userID),
		Name:                req.Name,
		PrimaryPersonsMode:  parsePrimaryPersonsMode(req.PrimaryPersonsMode),
		PrimaryPersonsCount: int32(req.PrimaryPersonsCount),
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("create event type failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]any{"event_type": toEventTypeJSON(resp.GetEventType())})
}

// GetEventType returns event type by ID.
// @Summary Get event type
// @Tags event-types
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param event_type_id path string true "Event type ID"
// @Success 200 {object} eventTypeSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/event-types/{event_type_id} [get]
func (h *Handler) GetEventType(w http.ResponseWriter, r *http.Request) {
	userID, err := middleware.UserIDFromContext(r.Context())
	if err != nil {
		response.Error(w, http.StatusUnauthorized, "unauthorized", "invalid token claims")
		return
	}

	eventTypeID := chi.URLParam(r, "event_type_id")
	if strings.TrimSpace(eventTypeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "event_type_id is required")
		return
	}

	resp, err := h.client.GetEventType(r.Context(), &eventspb.GetEventTypeRequest{
		RequestUserId: int32(userID),
		EventTypeId:   eventTypeID,
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("get event type failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]any{"event_type": toEventTypeJSON(resp.GetEventType())})
}

// ListEventTypes lists available system and custom event types.
// @Summary List event types
// @Tags event-types
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} eventTypesSuccessResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/event-types/ [get]
func (h *Handler) ListEventTypes(w http.ResponseWriter, r *http.Request) {
	userID, err := middleware.UserIDFromContext(r.Context())
	if err != nil {
		response.Error(w, http.StatusUnauthorized, "unauthorized", "invalid token claims")
		return
	}

	resp, err := h.client.ListEventTypes(r.Context(), &eventspb.ListEventTypesRequest{RequestUserId: int32(userID)})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("list event types failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	eventTypes := make([]map[string]any, 0, len(resp.GetEventTypes()))
	for _, eventType := range resp.GetEventTypes() {
		eventTypes = append(eventTypes, toEventTypeJSON(eventType))
	}

	response.OK(w, map[string]any{"event_types": eventTypes})
}

// DeleteEventType deletes a custom event type.
// @Summary Delete event type
// @Tags event-types
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param event_type_id path string true "Event type ID"
// @Success 200 {object} eventsStatusSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/event-types/{event_type_id} [delete]
func (h *Handler) DeleteEventType(w http.ResponseWriter, r *http.Request) {
	userID, err := middleware.UserIDFromContext(r.Context())
	if err != nil {
		response.Error(w, http.StatusUnauthorized, "unauthorized", "invalid token claims")
		return
	}

	eventTypeID := chi.URLParam(r, "event_type_id")
	if strings.TrimSpace(eventTypeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "event_type_id is required")
		return
	}

	err = h.client.DeleteEventType(r.Context(), &eventspb.DeleteEventTypeRequest{
		RequestUserId: int32(userID),
		EventTypeId:   eventTypeID,
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("delete event type failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]string{"status": "ok"})
}

// CreateEvent creates an event in a tree.
// @Summary Create event
// @Tags events
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param tree_id path string true "Tree ID"
// @Param request body upsertEventRequest true "Request body"
// @Success 200 {object} eventSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/events/{tree_id} [post]
func (h *Handler) CreateEvent(w http.ResponseWriter, r *http.Request) {
	var req upsertEventRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "bad_request", "invalid request body")
		return
	}

	treeID := chi.URLParam(r, "tree_id")
	if strings.TrimSpace(treeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "tree_id is required")
		return
	}

	resp, err := h.client.CreateEvent(r.Context(), &eventspb.CreateEventRequest{
		TreeId:              treeID,
		EventTypeId:         req.EventTypeID,
		PrimaryPersonIds:    req.PrimaryPersonIDs,
		AdditionalPersonIds: req.AdditionalPersonIDs,
		DateIso:             req.DateISO,
		DateUnknown:         req.DateUnknown,
		DatePrecision:       parseDatePrecision(req.DatePrecision),
		DateBound:           parseDateBound(req.DateBound),
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("create event failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]any{"event": toEventJSON(resp.GetEvent())})
}

// GetEvent returns event by ID from tree.
// @Summary Get event
// @Tags events
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param tree_id path string true "Tree ID"
// @Param event_id path string true "Event ID"
// @Success 200 {object} eventSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/events/{tree_id}/{event_id} [get]
func (h *Handler) GetEvent(w http.ResponseWriter, r *http.Request) {
	treeID := chi.URLParam(r, "tree_id")
	if strings.TrimSpace(treeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "tree_id is required")
		return
	}

	eventID := chi.URLParam(r, "event_id")
	if strings.TrimSpace(eventID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "event_id is required")
		return
	}

	resp, err := h.client.GetEvent(r.Context(), &eventspb.GetEventRequest{
		TreeId:  treeID,
		EventId: eventID,
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("get event failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]any{"event": toEventJSON(resp.GetEvent())})
}

// ListEventsByTree returns all events in a tree.
// @Summary List events by tree
// @Tags events
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param tree_id path string true "Tree ID"
// @Success 200 {object} eventsSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/events/{tree_id} [get]
func (h *Handler) ListEventsByTree(w http.ResponseWriter, r *http.Request) {
	treeID := chi.URLParam(r, "tree_id")
	if strings.TrimSpace(treeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "tree_id is required")
		return
	}

	resp, err := h.client.ListEventsByTree(r.Context(), &eventspb.ListEventsByTreeRequest{
		TreeId: treeID,
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("list events by tree failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	events := make([]map[string]any, 0, len(resp.GetEvents()))
	for _, event := range resp.GetEvents() {
		events = append(events, toEventJSON(event))
	}

	response.OK(w, map[string]any{"events": events})
}

// UpdateEvent updates an existing event.
// @Summary Update event
// @Tags events
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param tree_id path string true "Tree ID"
// @Param event_id path string true "Event ID"
// @Param request body upsertEventRequest true "Request body"
// @Success 200 {object} eventSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/events/{tree_id}/{event_id} [put]
func (h *Handler) UpdateEvent(w http.ResponseWriter, r *http.Request) {
	treeID := chi.URLParam(r, "tree_id")
	if strings.TrimSpace(treeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "tree_id is required")
		return
	}

	eventID := chi.URLParam(r, "event_id")
	if strings.TrimSpace(eventID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "event_id is required")
		return
	}

	var req upsertEventRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		response.Error(w, http.StatusBadRequest, "bad_request", "invalid request body")
		return
	}

	resp, err := h.client.UpdateEvent(r.Context(), &eventspb.UpdateEventRequest{
		TreeId:              treeID,
		EventId:             eventID,
		EventTypeId:         req.EventTypeID,
		PrimaryPersonIds:    req.PrimaryPersonIDs,
		AdditionalPersonIds: req.AdditionalPersonIDs,
		DateIso:             req.DateISO,
		DateUnknown:         req.DateUnknown,
		DatePrecision:       parseDatePrecision(req.DatePrecision),
		DateBound:           parseDateBound(req.DateBound),
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("update event failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]any{"event": toEventJSON(resp.GetEvent())})
}

// DeleteEvent deletes event by ID from tree.
// @Summary Delete event
// @Tags events
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param tree_id path string true "Tree ID"
// @Param event_id path string true "Event ID"
// @Success 200 {object} eventsStatusSuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 401 {object} response.ErrorResponse
// @Failure 403 {object} response.ErrorResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 409 {object} response.ErrorResponse
// @Failure 429 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /api/events/{tree_id}/{event_id} [delete]
func (h *Handler) DeleteEvent(w http.ResponseWriter, r *http.Request) {
	treeID := chi.URLParam(r, "tree_id")
	if strings.TrimSpace(treeID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "tree_id is required")
		return
	}

	eventID := chi.URLParam(r, "event_id")
	if strings.TrimSpace(eventID) == "" {
		response.Error(w, http.StatusBadRequest, "bad_request", "event_id is required")
		return
	}

	err := h.client.DeleteEvent(r.Context(), &eventspb.DeleteEventRequest{
		TreeId:  treeID,
		EventId: eventID,
	})
	if err != nil {
		status, msg := grpcerr.HTTPStatus(err)
		h.log.Error("delete event failed", slog.String("error", err.Error()))
		response.Error(w, status, "events_error", msg)
		return
	}

	response.OK(w, map[string]string{"status": "ok"})
}

func parsePrimaryPersonsMode(v string) eventspb.PrimaryPersonsMode {
	switch strings.ToUpper(strings.TrimSpace(v)) {
	case "FIXED":
		return eventspb.PrimaryPersonsMode_PRIMARY_PERSONS_MODE_FIXED
	case "UNLIMITED":
		return eventspb.PrimaryPersonsMode_PRIMARY_PERSONS_MODE_UNLIMITED
	default:
		return eventspb.PrimaryPersonsMode_PRIMARY_PERSONS_MODE_UNSPECIFIED
	}
}

func parseDatePrecision(v string) eventspb.EventDatePrecision {
	switch strings.ToUpper(strings.TrimSpace(v)) {
	case "DAY":
		return eventspb.EventDatePrecision_EVENT_DATE_PRECISION_DAY
	case "MONTH":
		return eventspb.EventDatePrecision_EVENT_DATE_PRECISION_MONTH
	case "YEAR":
		return eventspb.EventDatePrecision_EVENT_DATE_PRECISION_YEAR
	default:
		return eventspb.EventDatePrecision_EVENT_DATE_PRECISION_UNSPECIFIED
	}
}

func parseDateBound(v string) eventspb.EventDateBound {
	switch strings.ToUpper(strings.TrimSpace(v)) {
	case "EXACT":
		return eventspb.EventDateBound_EVENT_DATE_BOUND_EXACT
	case "NOT_BEFORE":
		return eventspb.EventDateBound_EVENT_DATE_BOUND_NOT_BEFORE
	case "NOT_AFTER":
		return eventspb.EventDateBound_EVENT_DATE_BOUND_NOT_AFTER
	default:
		return eventspb.EventDateBound_EVENT_DATE_BOUND_UNSPECIFIED
	}
}

func toEventTypeJSON(eventType *eventspb.EventType) map[string]any {
	if eventType == nil {
		return map[string]any{}
	}

	return map[string]any{
		"id":                    eventType.GetId(),
		"owner_user_id":         eventType.GetOwnerUserId(),
		"is_system":             eventType.GetIsSystem(),
		"name":                  eventType.GetName(),
		"primary_persons_mode":  eventType.GetPrimaryPersonsMode().String(),
		"primary_persons_count": eventType.GetPrimaryPersonsCount(),
	}
}

func toEventJSON(event *eventspb.Event) map[string]any {
	if event == nil {
		return map[string]any{}
	}

	return map[string]any{
		"id":                    event.GetId(),
		"tree_id":               event.GetTreeId(),
		"event_type_id":         event.GetEventTypeId(),
		"primary_person_ids":    event.GetPrimaryPersonIds(),
		"additional_person_ids": event.GetAdditionalPersonIds(),
		"date_iso":              event.GetDateIso(),
		"date_unknown":          event.GetDateUnknown(),
		"is_autogenerated":      event.GetIsAutogenerated(),
		"date_precision":        event.GetDatePrecision().String(),
		"date_bound":            event.GetDateBound().String(),
		"created_at_unix":       event.GetCreatedAtUnix(),
		"updated_at_unix":       event.GetUpdatedAtUnix(),
	}
}